# bank/src/test/resources/spec/digitalbank/openapi/DigitalBank-OpenAPI3.0.json
openapi: 3.0.0
info:
  description: Digial Bank API provides Administration and User functionality through API
    Endpoints.
  version: API TOS
  title: Digital Bank API
  termsOfService: Terms of Service
  contact:
    name: Digital Bank
    url: www.demo.io
    email: digitalbank@demo.io
  license:
    name: License of API
    url: API License URL
...
paths:
...
  /api/v1/user:
    get:
      tags:
        - user-controller
      summary: getUser
      operationId: getUserUsingGET_1
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
        - Bearer_token: []
      deprecated: false
    post:
      tags:
        - user-controller
      summary: createNewUser
      operationId: createNewUserUsingPOST
      parameters:
        - name: role
          in: query
          description: role
          required: false
          schema:
            type: string
            default: USER
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewUser"
        description: newUser
        required: true
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
        "201":
          description: Created
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
        - Bearer_token: []
      deprecated: false
...
  /api/v1/user/password:
    put:
      tags:
        - user-controller
      summary: changePassword
      operationId: changePasswordUsingPUT
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePassword"
        description: changePassword
        required: true
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
        "201":
          description: Created
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
        - Bearer_token: []
      deprecated: false
  /api/v1/user/profile:
    get:
      tags:
        - user-controller
      summary: getUserProfile
      operationId: getUserProfileUsingGET
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
        - Bearer_token: []
      deprecated: false
    put:
      tags:
        - user-controller
      summary: updateUserProfile
      operationId: updateUserProfileUsingPUT
      requestBody:
        $ref: "#/components/requestBodies/UpdateUser"
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
        "201":
          description: Created
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
        - Bearer_token: []
      deprecated: false
  /api/v1/user/role:
    get:
      tags:
        - user-controller
      summary: getAuthority
      operationId: getAuthorityUsingGET
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
        - Bearer_token: []
      deprecated: false
  "/api/v1/user/{id}":
    get:
      tags:
        - user-controller
      summary: getUser
      operationId: getUserUsingGET
      parameters:
        - name: id
          in: path
          description: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
        - Bearer_token: []
      deprecated: false
    delete:
      tags:
        - user-controller
      summary: deleteUser
      operationId: deleteUserUsingDELETE
      parameters:
        - name: id
          in: path
          description: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
        "204":
          description: No Content
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
      security:
        - Bearer_token: []
      deprecated: false
...
  "/api/v1/user/{id}/password":
    put:
      tags:
        - user-controller
      summary: setPassword
      operationId: setPasswordUsingPUT
      parameters:
        - name: id
          in: path
          description: id
          required: true
          schema:
            type: integer
            format: int64
        - name: newPassword
          in: query
          description: newPassword
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
        "201":
          description: Created
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
        - Bearer_token: []
      deprecated: false
  "/api/v1/user/{id}/password/unexpire":
    put:
      tags:
        - user-controller
      summary: setPasswordUnexpired
      operationId: setPasswordUnexpiredUsingPUT
      parameters:
        - name: id
          in: path
          description: id
          required: true
          schema:
            type: integer
            format: int64
        - name: unexpire
          in: query
          description: unexpire
          required: true
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
        "201":
          description: Created
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
        - Bearer_token: []
      deprecated: false
  "/api/v1/user/{id}/profile":
    get:
      tags:
        - user-controller
      summary: getUserProfile
      operationId: getUserProfileUsingGET_1
      parameters:
        - name: id
          in: path
          description: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
        - Bearer_token: []
      deprecated: false
    put:
      tags:
        - user-controller
      summary: updateUserProfile
      operationId: updateUserProfileUsingPUT_1
      parameters:
        - name: id
          in: path
          description: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        $ref: "#/components/requestBodies/UpdateUser"
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
        "201":
          description: Created
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
        - Bearer_token: []
      deprecated: false
  "/api/v1/user/{id}/role":
    get:
      tags:
        - user-controller
      summary: getAuthority
      operationId: getAuthorityUsingGET_1
      parameters:
        - name: id
          in: path
          description: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
        - Bearer_token: []
      deprecated: false
    put:
      tags:
        - user-controller
      summary: addAuthority
      operationId: addAuthorityUsingPUT
      parameters:
        - name: id
          in: path
          description: id
          required: true
          schema:
            type: integer
            format: int64
        - name: role
          in: query
          description: role
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
        "201":
          description: Created
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
        - Bearer_token: []
      deprecated: false
    delete:
      tags:
        - user-controller
      summary: removeAuthority
      operationId: removeAuthorityUsingDELETE
      parameters:
        - name: id
          in: path
          description: id
          required: true
          schema:
            type: integer
            format: int64
        - name: role
          in: query
          description: role
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
        "204":
          description: No Content
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
      security:
        - Bearer_token: []
      deprecated: false
  "/api/v1/user/{id}/state/enable":
    put:
      tags:
        - user-controller
      summary: setUserEnabled
      operationId: setUserEnabledUsingPUT
      parameters:
        - name: enabled
          in: query
          description: enabled
          required: true
          schema:
            type: boolean
        - name: id
          in: path
          description: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
        "201":
          description: Created
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
        - Bearer_token: []
      deprecated: false
  "/api/v1/user/{id}/state/unexpire":
    put:
      tags:
        - user-controller
      summary: setUserUnexpired
      operationId: setUserUnexpiredUsingPUT
      parameters:
        - name: id
          in: path
          description: id
          required: true
          schema:
            type: integer
            format: int64
        - name: unexpire
          in: query
          description: unexpire
          required: true
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
        "201":
          description: Created
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
        - Bearer_token: []
      deprecated: false
  "/api/v1/user/{id}/state/unlock":
    put:
      tags:
        - user-controller
      summary: setUserUnlocked
      operationId: setUserUnlockedUsingPUT
      parameters:
        - name: id
          in: path
          description: id
          required: true
          schema:
            type: integer
            format: int64
        - name: unlock
          in: query
          description: unlock
          required: true
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
        "201":
          description: Created
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
        - Bearer_token: []
      deprecated: false
  /api/v1/users:
    get:
      tags:
        - user-controller
      summary: getAllUsers
      operationId: getAllUsersUsingGET
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
        - Bearer_token: []
      deprecated: false
servers:
  - url: //localhost:8080/bank
components:
  requestBodies:
    UpdateUser:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/UpdateUser"
      description: profile
      required: true
...
  securitySchemes:
    Bearer_token:
      type: apiKey
      name: Authorization
      in: header
  schemas:
...
    ChangePassword:
      type: object
      properties:
        currentPasword:
          type: string
        newPassword:
          type: string
      title: ChangePassword
...
    NewUser:
      type: object
      properties:
        address:
          type: string
        country:
          type: string
        dob:
          type: string
          example: MM/dd/yyyy
        emailAddress:
          type: string
        firstName:
          type: string
        gender:
          type: string
        homePhone:
          type: string
        lastName:
          type: string
        locality:
          type: string
        mobilePhone:
          type: string
        password:
          type: string
        postalCode:
          type: string
        region:
          type: string
        ssn:
          type: string
        title:
          type: string
        workPhone:
          type: string
      title: NewUser
...
    UpdateUser:
      type: object
      properties:
        address:
          type: string
        country:
          type: string
        firstName:
          type: string
        gender:
          type: string
        homePhone:
          type: string
        lastName:
          type: string
        locality:
          type: string
        mobilePhone:
          type: string
        postalCode:
          type: string
        region:
          type: string
        title:
          type: string
        workPhone:
          type: string
      title: UpdateUser